#116 Prepare for versioned release
