+++ doc.go
@@ -1,0 +1,4 @@
+/*
+The Senzing go-helpers packages contain utility routines for doing common tasks.
+*/
+package main
