+++ fileutil/doc.go
@@ -1,1 +1,4 @@
+/*
+Package fileutil containes functions for managing files.
+*/
 package fileutil
