+++ record/doc.go
@@ -1,5 +1,6 @@
 /*
-The `record` package provides a validating data type for Senzing records.
+Package record is used to validate Senzing records.
+
 The Record structure can be used to parse JSON into records.
 There are functions for validating strings and creating Records.
 */
