+++ settings/doc.go
@@ -1,4 +1,4 @@
 /*
-The settings package is used to create the JSON needed for initializing the Senzing runtime objects.
+Package settings is used to create the JSON needed for Senzing runtime object settings.
 */
 package settings
