+++ testfixtures/doc.go
@@ -1,4 +1,4 @@
 /*
-testfixtures is a set of artifacts used during "go test".
+Package testfixtures is a set of artifacts used during "go test".
 */
 package testfixtures
