+++ truthset/doc.go
@@ -1,0 +1,4 @@
+/*
+Package truthset contains a Go representation of the Senzing truthset.
+*/
+package truthset
